Split MainActivity into two peers, old and new to keep logic isolated.
This change updates MainActivity to defer lifecycle callbacks and other
relevant methods to which ever peer is active, old or new.

The old peer implements the logic for the old fragments.
The new peer implements the logic for the new fragments.
MainActivity implements the logic that is shared or common between the two like
last tab, post call, search, dialpad, ect.

Bug: 72525324
Test: pending
PiperOrigin-RevId: 184317828
Change-Id: Ie73733f2a3837c9d63e54cf5b142984633340731