Check if the activity is still alive before refreshing unread tab counts.
My hypothesis is that there is a race condition where the call log changes and
we are leaving the activity before the content observer is unregistered or the
loader it starts is finished. This results in a callback calling #fetchCalls
while Dialer is in the background, resulting in a crash. This change will
confirm my hypothesis.

Even if my hypothesis is incorrect, we have similar checks throughout this
class which indicates to me that some issue exists here and that this check is
safe to do.

Bug: 73972084
Test: n/a
PiperOrigin-RevId: 187700318
Change-Id: I98cea86b5b7041c4f4aac09544445aaade6bfd40