Implement non-ui logic for delete/undo delete voicemails
When a voicemail is deleted we have to allow the user the option of cancelling the deleting of voicemail. This is done by showing the user a snack bar which when a user taps will prevent the voicemail from being deleted.
This CL only adds the underlying logic for the delete/undelete to take place via a ScheduledFuture, it does not fix all the UI issues that come with it, which will be sent in a follow up CL, issues such as:
-Updating the group label when a delete/undo happens
-Updating the entries above and below when a delete/undo happens
-Showing the empty view immediately when a user deletes a single voicemail
-Removing the empty view immediately when a user taps undo
-Updating tests

Bug: 69858266
Test: Manual
PiperOrigin-RevId: 188424489
Change-Id: Iecf5cb7be5b80fa5e91acaad3a7cc4a6c8702521