Use SwitchPreferenceWithClickableSummary instead of SwitchPreference for VM Donation
The learn more link needs to be in the summary of the switch preference, however having a hyperlink in the summary text view of a switch preference does not open the hyperlink. To do this we made our custom switch preference which allows the summary to be clicked. All other aspects of the switch preference remain the same.

Bug: 74033229
Test: Unit test
PiperOrigin-RevId: 189978676
Change-Id: I31e744f3545e576ee3f5ac4a8fee249e22835e19