Support new call log fragment in old peer.
This makes the old peer read the CallLogConfig#isNewCallLogFragmentEnabled and show the old or new fragment accordingly.

If the user is viewing the NewCallLog and the CallLogConfig needs to disable the framework, the new fragment is immediately replaced with the old one. This is necessary because if the user were to scroll the fragment, the AnnotatedCallLog and PhoneLookupHistory databases would be read, which would trigger creation.

I don't expect this to be a common case because 1) we hopefully never have to disable the framework and 2) Framework is only updated on Phenotype broadcasts and JobScheduler jobs, which hopefully don't typically happen when user is viewing the call log. However, I still want to make sure that if it happens we don't irreversibly break users when we turn the framework back on.

I tested this by flipping flags and observing underlying data being removed:

> dialer-cmd configprovider set new_voicemail_fragment_enabled false

> adb shell ls /data/data/com.google.android.dialer/databases/ && echo && adb shell cat /data/user_de/0/com.google.android.dialer/shared_prefs/com.google.android.dialer_preferences.xml

I test flipping flags back and forth on the call log tab, speed dial tab, and while the activity was paused (pressing Home after viewing call log).

Note that this CL doesn't address showing missed calls and badge counts correctly with the new fragment; that will come in a later CL.

Bug: 74821995
Test: unit and manual
PiperOrigin-RevId: 190706481
Change-Id: I618d9c1649169abd65733502cfebc662a835e787