Support new voicemail fragment in old peer.
This makes the old peer read the CallLogConfig#isNewVoicemailFragmentEnabled and show the old or new fragment accordingly.

If the user is viewing the NewVoicemail and the CallLogConfig needs to disable the framework, the new fragment is immediately replaced with the old one. This is necessary because if the user were to scroll the fragment, the AnnotatedCallLog database would be read, which would trigger creation.

I tested this by flipping flags and observing underlying data being removed:

> dialer-cmd configprovider set new_voicemail_fragment_enabled false

> adb shell ls /data/data/com.google.android.dialer/databases/ && echo && adb shell cat /data/user_de/0/com.google.android.dialer/shared_prefs/com.google.android.dialer_preferences.xml

I test flipping flags back and forth on the voicemail tab, call log tab and ensuring that they are independent.

Bug: 77601968
Test: unit and manual. Some tests are failing, so to ensure we can have the voicemail ready for the bug bash tomorrow, I've ignored them temporarily but will be fixed in a follow up CL (tracked by b/77601893)
PiperOrigin-RevId: 191738860
Change-Id: I24ca38b862e98324cf802a3020e7e9df31c0b966