Solve BiDi problem when phone number is concatenated with RTL-string
Solves BiDi problem that might cause phone number to be displayed
incorrectly in the dialer. Issue might occur when a phone number
is concatenated with a custom number label containing an RTL-string.

Solved by changing affected labels from TextView to BidiTextView.

Removed function spanDisplayNumber in BottomRow which are now obsolete.

Updates the DialerBidiFormatter to prevent whitespace to be misplaced
within a BidiTextView.

BUG=78464687
TEST=Manual

Public-Origin-Change-Id: I789a6f76a5ca805c03a437813ce627eef7a0bf2e
Author: Torbjorn Eklund <torbjorn.eklund@sony.com>
Bug: 78464687
Test: Manual
PiperOrigin-RevId: 198823629
Change-Id: Id35c3ebf514609b4a1fe10a7ae5a297bde7c7456