Removed "last_tab_enabled" flag. END_PUBLIC
The bottom nav buttons do not work correctly when this flag is not enabled (see bugs). This flag is intended to always be true via an experiment but can be false for example when the user clears data or just launched dialer for the first time and hasn't received config flags.

It is not likely that we will ever want to turn off this behavior, so this change simply removes all support for it.

TEST=unit
Bug: 109762667,79485008
Test: unit
PiperOrigin-RevId: 199660314
Change-Id: I105f6c553c7541673b6d5dd7abf4d29f566e155f